Hide report pages behind GCS config
