+++ web/app/history/page.tsx
@@ -1,6 +1,8 @@
 import type { Metadata } from "next";
 import Link from "next/link";
 import HistoryContent from "./HistoryContent";
+import ReportPageUnavailable from "@/components/ReportPageUnavailable";
+import { hasReportPagesEnabled } from "@/lib/report-pages";
 import { APP_NAME } from "@/lib/pwa-metadata";
 
 export const metadata: Metadata = {
@@ -10,15 +12,23 @@ export const metadata: Metadata = {
 };
 
 export default function HistoryPage() {
+  if (!hasReportPagesEnabled()) {
+    return (
+      <ReportPageUnavailable
+        title="Исторически данни"
+        message="Тази страница е скрита, защото GCS_GENERIC_BUCKET не е конфигуриран за този fork."
+        backHref="/sources"
+        backLabel="Източници"
+      />
+    );
+  }
+
   return (
     <div className="min-h-screen bg-neutral-light flex flex-col">
       {/* Page header */}
       <div className="max-w-7xl mx-auto w-full px-4 sm:px-6 lg:px-8 py-12">
         <div className="mb-6">
-          <Link
-            href="/"
-            className="text-primary hover:text-primary-hover inline-flex items-center gap-2"
-          >
+          <Link href="/" className="text-primary hover:text-primary-hover inline-flex items-center gap-2">
             <span>←</span>
             <span>Начало</span>
           </Link>
