+++ web/lib/report-pages.ts
@@ -1,0 +1,3 @@
+export function hasReportPagesEnabled() {
+  return Boolean(process.env.GCS_GENERIC_BUCKET?.trim());
+}
